+++ packages/ZbayLite/src/renderer/containers/widgets/channels/ChannelMessage.js
@@ -2,11 +2,11 @@ import { connect } from 'react-redux'
 import { bindActionCreators } from 'redux'
 
 import ChannelMessage from '../../../components/widgets/channels/ChannelMessage'
-import pendingMessagesHandlers from '../../../store/handlers/pendingMessages'
+import operationsHandlers from '../../../store/handlers/operations'
 import channelHandlers from '../../../store/handlers/channel'
 
 export const mapDispatchToProps = (dispatch, ownProps) => bindActionCreators({
-  onCancel: () => pendingMessagesHandlers.actions.removeMessage(ownProps.message.get('id')),
+  onCancel: () => operationsHandlers.actions.removeOperation(ownProps.message.get('id')),
   onResend: () => channelHandlers.epics.resendMessage(ownProps.message.toJS())
 }, dispatch)