+++ packages/ZbayLite/src/renderer/store/create.js
@@ -1,6 +1,7 @@
 import { applyMiddleware, createStore, compose } from 'redux'
 import thunk from 'redux-thunk'
 import promise from 'redux-promise-middleware'
+import createDebounce from 'redux-debounced'
 
 import reducers from './reducers'
 import { errorsMiddleware } from './middlewares'
@@ -18,5 +19,5 @@ export default ({
 } = {}) => createStore(
   reducers,
   initialState,
-  composer(applyMiddleware(...[errorsMiddleware, thunk, promise()]))
+  composer(applyMiddleware(...[errorsMiddleware, createDebounce(), thunk, promise()]))
 )