+++ packages/ZbayLite/src/renderer/store/handlers/messagesQueue.js
@@ -1,0 +1,116 @@
+import Immutable from 'immutable'
+import * as R from 'ramda'
+import crypto from 'crypto'
+import { createAction, handleActions } from 'redux-actions'
+
+import selectors from '../selectors/messagesQueue'
+import channelsSelectors from '../selectors/channels'
+import { messageToTransfer } from '../../zbay/messages'
+import operationsHandlers, { PendingMessageOp, operationTypes } from './operations'
+import notificationsHandlers from './notifications'
+import { errorNotification } from './utils'
+import { getClient } from '../../zcash'
+
+export const DEFAULT_DEBOUNCE_INTERVAL = 3000
+
+export const PendingMessage = Immutable.Record({
+  channelId: '',
+  message: null
+}, 'PendingMessage')
+
+export const initialState = Immutable.Map()
+
+const addMessage = createAction('ADD_PENDING_MESSAGE', ({ message, channelId }) => {
+  const messageDigest = crypto.createHash('sha256')
+  const messageEssentials = R.pick(['type', 'sender'])(message)
+  return {
+    key: messageDigest.update(
+      JSON.stringify({
+        ...messageEssentials,
+        channelId
+      })
+    ).digest('hex'),
+    message,
+    channelId
+  }
+})
+const removeMessage = createAction('REMOVE_PENDING_MESSAGE')
+
+export const actions = {
+  addMessage,
+  removeMessage
+}
+
+const _sendPendingMessages = async (dispatch, getState) => {
+  const messages = selectors.queue(getState())
+  await Promise.all(
+    messages.map(async (msg, key) => {
+      const channel = channelsSelectors.channelById(msg.channelId)(getState())
+      const transfer = await messageToTransfer({
+        message: msg.message.toJS(),
+        channel: channel.toJS()
+      })
+
+      let opId
+      try {
+        opId = await getClient().payment.send(transfer)
+      } catch (err) {
+        dispatch(notificationsHandlers.actions.enqueueSnackbar(
+          errorNotification({
+            message: 'Couldn\'t send the message, please check node connection.'
+          })
+        ))
+        return
+      }
+      dispatch(removeMessage(key))
+      await dispatch(operationsHandlers.epics.observeOperation({
+        opId,
+        type: operationTypes.pendingMessage,
+        meta: PendingMessageOp({
+          channelId: channel.get('id'),
+          message: msg.message
+        })
+      }))
+    }).values()
+  )
+}
+
+const sendPendingMessages = () => {
+  const thunk = _sendPendingMessages
+  thunk.meta = {
+    debounce: {
+      time: process.env.ZBAY_DEBOUNCE_MESSAGE_INTERVAL || DEFAULT_DEBOUNCE_INTERVAL,
+      key: 'SEND_PENDING_MESSAGES'
+    }
+  }
+  return thunk
+}
+
+const addMessageEpic = (payload) => async (dispatch) => {
+  dispatch(addMessage(payload))
+  await dispatch(sendPendingMessages())
+}
+
+export const epics = {
+  sendPendingMessages,
+  addMessage: addMessageEpic
+}
+
+export const reducer = handleActions({
+  [addMessage]: (state, { payload: { channelId, message, key } }) => {
+    if (state.has(key)) {
+      return state.updateIn([key, 'message', 'message'], m => [m, message.message].join('\n'))
+    }
+    return state.set(key, PendingMessage({
+      channelId,
+      message: Immutable.fromJS(message)
+    }))
+  },
+  [removeMessage]: (state, { payload: key }) => state.remove(key)
+}, initialState)
+
+export default {
+  actions,
+  epics,
+  reducer
+}