+++ packages/ZbayLite/src/renderer/store/selectors/pendingMessages.js
@@ -1,9 +1,0 @@
-import { createSelector } from 'reselect'
-
-const store = s => s
-
-const pendingMessages = createSelector(store, state => state.get('pendingMessages'))
-
-export default {
-  pendingMessages
-}